+++ stigg-java-core/src/test/kotlin/io/stigg/models/internal_/beta/eventqueues/EventQueueListParamsTest.kt
@@ -1,0 +1,13 @@
+// File generated from our OpenAPI spec by Stainless.
+
+package io.stigg.models.internal_.beta.eventqueues
+
+import org.junit.jupiter.api.Test
+
+internal class EventQueueListParamsTest {
+
+    @Test
+    fun create() {
+        EventQueueListParams.builder().build()
+    }
+}