+++ stigg-java-proguard-test/src/test/kotlin/io/stigg/proguard/ProGuardCompatibilityTest.kt
@@ -50,6 +50,7 @@ internal class ProGuardCompatibilityTest {
 
         assertThat(client).isNotNull()
         assertThat(client.v1()).isNotNull()
+        assertThat(client.internal_()).isNotNull()
     }
 
     @Test